[BE] Projects CRUD / apply logger + exception handling ***     project.controller: exception을 리턴하도록 수정     project.service: exception handling 추가     project.module: CUD에 UnAuthMiddleware 적용     un-auth.middleware: 비로그인 상태일 경우 접근 제한하는 미들웨어 ***

+++ server/src/project/project.service.ts
@@ -1,6 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  InternalServerErrorException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import * as fs from 'fs';
+import { LoggerService } from 'src/logger/logger.service';
 import { Repository } from 'typeorm';
 import { Project } from './project.entity';
 
@@ -9,21 +15,38 @@ export class ProjectService {
   constructor(
     @InjectRepository(Project)
     private readonly projectRepository: Repository<Project>,
+    private readonly logger: LoggerService,
   ) {
     const FOLDER_PROJECTS = `${__dirname}/../../markdowns/projects`;
     !fs.existsSync(FOLDER_PROJECTS) && fs.mkdirSync(FOLDER_PROJECTS); // projects 폴더가 존재하지 않을 경우 생성
   }
 
   async getProjects(): Promise<Project[]> {
-    const projects: Project[] = await this.projectRepository.find({
-      select: ['id', 'subject', 'writer'],
-    });
+    try {
+      const projects: Project[] = await this.projectRepository.find({
+        select: ['id', 'subject', 'writer'],
+      });
 
-    return projects;
+      return projects;
+    } catch (err) {
+      this.logger.error(err);
+      throw new InternalServerErrorException();
+    }
   }
 
   async getProject(id: number) {
     const project: Project = await this.projectRepository.findOne(id);
+    if (!project) {
+      this.logger.error(`Cannot found project id ${id}`);
+      throw new HttpException(
+        {
+          status: HttpStatus.NOT_FOUND,
+          error: `Cannot found project id ${id}`,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     const { subject, file_path } = project;
 
     const markdown = fs.readFileSync(file_path, 'utf8');
@@ -31,15 +54,20 @@ export class ProjectService {
   }
 
   createProject(subject: string, markdown: string) {
-    const file_path = this.writeCodeIntoFile(subject, markdown);
+    try {
+      const file_path = this.writeCodeIntoFile(subject, markdown);
 
-    const project: Project = {
-      subject,
-      filename: subject,
-      file_path,
-    };
+      const project: Project = {
+        subject,
+        filename: subject,
+        file_path,
+      };
 
-    this.projectRepository.save(project);
+      this.projectRepository.save(project);
+    } catch (err) {
+      this.logger.error(err);
+      throw new InternalServerErrorException();
+    }
   }
 
   writeCodeIntoFile(subject: string, markdown: string) {
@@ -48,15 +76,26 @@ export class ProjectService {
       fs.writeFileSync(file_path, markdown);
       return file_path;
     } catch (err) {
-      console.error(err);
+      throw new Error(err);
     }
   }
 
   async updateProject(id: number, subject: string, markdonw: string) {
-    try {
-      const fileToUpdate: Project = await this.projectRepository.findOne(id);
-      let { file_path } = fileToUpdate;
+    const fileToUpdate: Project = await this.projectRepository.findOne(id);
+    if (!fileToUpdate) {
+      this.logger.error(`Cannot found project id ${id}`);
+      throw new HttpException(
+        {
+          status: HttpStatus.NOT_FOUND,
+          error: `Cannot found project id ${id}`,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
 
+    let { file_path } = fileToUpdate;
+
+    try {
       if (subject.length) {
         const new_file_path = `${__dirname}/../../markdowns/projects/${subject}.md`;
 
@@ -71,19 +110,32 @@ export class ProjectService {
 
       fs.writeFileSync(file_path, markdonw);
     } catch (err) {
-      console.log(err);
+      this.logger.error(err);
+      throw new InternalServerErrorException();
     }
   }
 
   async deleteProject(id: number) {
-    try {
-      const fileToDelete: Project = await this.projectRepository.findOne(id);
-      const { file_path } = fileToDelete;
+    const fileToDelete: Project = await this.projectRepository.findOne(id);
+    if (!fileToDelete) {
+      this.logger.error(`error: Cannot found project id ${id}`);
+      throw new HttpException(
+        {
+          status: HttpStatus.NOT_FOUND,
+          error: `Cannot found project id ${id}`,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
 
+    const { file_path } = fileToDelete;
+
+    try {
       fs.unlinkSync(file_path);
       await this.projectRepository.remove(fileToDelete);
     } catch (err) {
-      console.error(err);
+      this.logger.error(err);
+      throw new InternalServerErrorException();
     }
   }
 }
